requirements.txt
openai
langchain-openai
requests
certifi
langchain-community
python-dotenv
pip_system_certs
langchain
langchain-core
requests_oauthlib